delete fragment.rs
