+++ ucp/src/lib.rs
@@ -12,7 +12,6 @@ use tokio::{
 
 pub(crate) mod conn;
 pub(crate) mod cubic;
-pub(crate) mod fragment;
 pub mod packets;
 pub(crate) mod receive;
 pub(crate) mod send;
