+++ ucp/src/packets.rs
@@ -1,7 +1,5 @@
 use packet_derive::*;
 
-use crate::fragment::{FragmentHeader, FRAGMENT_FLAG};
-
 #[repr(u8)]
 #[derive(Clone, Copy)]
 pub enum Reliability {
@@ -126,4 +124,13 @@ impl Frame {
         }
         ret
     }
+}
+
+pub(crate) const FRAGMENT_FLAG: u8 = 0x10;
+
+#[derive(Clone)]
+pub(crate) struct FragmentHeader {
+    pub size: u32,
+    pub id: u16,
+    pub index: u32,
 }
